Restructure search tool utils and config layout
