+++ examples/data_preprocess/config/search_r1_like_data.yaml
